+++ code/parachain/runtime/picasso/src/xcmp.rs
@@ -1,7 +1,4 @@
-//! Setup of XCMP for parachain to allow cross chain transfers and other operations.
-//! Very similar to https://github.com/galacticcouncil/Basilisk-node/blob/master/runtime/basilisk/src/xcm.rs
-#![allow(unused_imports)] // allow until v2 xcm released (instead creating 2 runtimes)
-use super::*; // recursive dependency onto runtime
+use super::*;
 use codec::{Decode, Encode};
 use common::{
 	topology::{self},
@@ -210,20 +207,7 @@ impl xcm_executor::Config for XcmConfig {
 	type AssetTrap = CaptureAssetTrap;
 }
 
-parameter_type_with_key! {
-	pub OutgoingParachainMinFee: |location: MultiLocation| -> Option<Balance> {
-		#[allow(clippy::match_ref_pats)] // false positive
-		#[allow(clippy::match_single_binding)]
-		match (location.parents, location.first_interior()) {
-			(1, None) => Some(400_000_000_000),
-			(1, Some(Parachain(id)))  =>  {
-				let location = XcmAssetLocation::new(location.clone());
-				AssetsRegistry::min_xcm_fee(ParaId::from(*id), location).or(Some(u128::MAX))
-			},
-			_ => Some(u128::MAX),
-		}
-	};
-}
+
 
 impl orml_xtokens::Config for Runtime {
 	type Event = Event;
